requirements.txt
numpy~=1.18.1
PyYAML~=5.3.1
pandas~=1.2.1
pyarrow~=3.0.0